Finished levelling system, started the work on after game pop ups
Modified Levelsettings class. Now it is easy to extend and add infinite amount of levels. Added animated object interface, to control animations and stop them when player dies and when game ends. finished transition of scenes to levels and to menu, started the work on pop up windows after game, that display messages and score, plus lets you to go back to main menu or continue playing